Fixes for running rack-based applications.
Ruby changes:
- Put ESPIPE class in Errno module; it was defined at top-level which is
  incorrect, and Sinatra's error-reporting depends on it.
- Remove 'hacks' from irake; it doesn't exist anymore

IronRuby.Rack changes:
- Make sure request.body is always a string. Add Microsoft.Dynamic.dll to
  IronRuby.Rack. Update permissions instructions in README. Make sure the Ruby
  errors aren't swallowed because the error handler throws. Only try to delete
  the "perf" directory if it exists.
- Generate the rack env with Ruby types the first time around, and only execute
  Ruby code where needed.
- Make sure AppPath is a full path
- Remove locks in HttpHandler and HttpHandlerFactory, since ASP.NET does not
  do any thread-based load-balancing automatically.
- Requests are proessed with HttpAsyncHandler, which schedules each request in
  the ThreadPool, unblocking the thread listening for requests. Need to make
  sure this still processes requests in parallel.
- Lazy load web.config settings, and allow setting RACK_ENV from web.config.
- rack.input should never have a null contents (eg. StringIO.new(nil)).
- rack.version should be set to Rack::VERSION, rather than hardcoded.
- Manually set TOPLEVEL_BINDING since it's not set automatically when IronRuby
  is embedded.
- Fully expand the path of files added to the load path, so Ruby files always
  show "/" as the path separator.
- Always log exceptions to ironrack.log, and only show them to the HTML page
  when RACK_ENV is 'development'.
- Clean up little perf test, and add an actual test of IronRuby.Rack, though
  testing is still very lacking.
- Update Rails example app, and run all examples in RACK_ENV production mode.
- better command-line opts for ironruby.rack/test.rb, and cleaner output for
  ironruby.rack/test.bat

+++ Merlin/Main/Hosts/IronRuby.Rack/test.bat
@@ -1,30 +1,36 @@
 @echo off
 
-echo Running tests:
-
-set GEM_PATH=C:\Ruby\lib\ruby\gems\1.8
-set GEM_HOME=C:\Ruby\lib\ruby\gems\1.8
-rmdir /S /Q %~dp0perf
+if exist %~dp0perf rmdir /S /Q %~dp0perf
 mkdir %~dp0perf
 pushd %~dp0perf
 
-echo IronRuby Rack ...
-%~dp0..\..\bin\release\ir.exe %~dp0test.rb rack > ironruby-rack.txt
+set RUBY=%~dp0..\..\bin\release\ir.exe
 
-echo IronRuby Sinatra ...
-%~dp0..\..\bin\release\ir.exe %~dp0test.rb sinatra > ironruby-sinatra.txt
+<nul (set/p z=IronRuby Rack ... )
+%RUBY% %~dp0test.rb rack 1> ironruby-rack.txt 2>&1
+if "%ERRORLEVEL%" equ "0" ( echo [pass] ) else ( echo [fail] )
 
-echo IronRuby Rails ...
-%~dp0..\..\bin\release\ir.exe %~dp0test.rb rails > ironruby-rails.txt
+<nul (set/p z=IronRuby Sinatra ... )
+%RUBY% %~dp0test.rb sinatra 1> ironruby-sinatra.txt 2>&1
+if "%ERRORLEVEL%" equ "0" ( echo [pass] ) else ( echo [fail] )
 
-echo MRI Rack ...
-ruby %~dp0test.rb rack > ruby-rack.txt
+<nul (set/p z=IronRuby Rails ... )
+%RUBY% %~dp0test.rb rails 1> ironruby-rails.txt 2>&1
+if "%ERRORLEVEL%" equ "0" ( echo [pass] ) else ( echo [fail] )
 
-echo MRI Sinatra ...
-ruby %~dp0test.rb sinatra > ruby-sinatra.txt
+set RUBY=C:\Ruby\bin\ruby.exe
 
-echo MRI Rails ...
-ruby %~dp0test.rb rails > ruby-rails.txt
+<nul (set/p z=MRI Rack ... )
+%RUBY% %~dp0test.rb rack 1> ruby-rack.txt 2>&1
+if "%ERRORLEVEL%" equ "0" ( echo [pass] ) else ( echo [fail] )
+
+<nul (set/p z=MRI Sinatra ... )
+%RUBY% %~dp0test.rb sinatra 1> ruby-sinatra.txt 2>&1
+if "%ERRORLEVEL%" equ "0" ( echo [pass] ) else ( echo [fail] )
+
+<nul (set/p z=MRI Rails ... )
+%RUBY% %~dp0test.rb rails 1> ruby-rails.txt 2>&1
+if "%ERRORLEVEL%" equ "0" ( echo [pass] ) else ( echo [fail] )
 
 echo DONE. See the "perf" directory for results.
 popd
